fix(booking): restore review-to-payment navigation after wizard refactor (#276)
Keep review-step progression available after submission/resume so members with larger guest counts can return to payment, and add regression coverage for four-guest flows.
