+++ src/features/sections/components/__tests__/EventBookingWizard.test.tsx
@@ -371,6 +371,215 @@ describe("EventBookingWizard", () => {
     expect(screen.getByText(/may require organiser review/i)).toBeInTheDocument();
     expect(screen.getByText(/Alex Guest · Dietary: Vegetarian/i)).toBeInTheDocument();
     expect(screen.getByText(/Sam Extra · Dietary: Gluten free/i)).toBeInTheDocument();
+    expect(screen.getByRole("button", { name: "Confirm booking" })).toBeInTheDocument();
+  });
+
+  it("shows Confirm booking on review when four guests are requested", async () => {
+    const user = userEvent.setup();
+    vi.mocked(reactGenerated.useGetMyBookingsForEvent).mockReturnValue({
+      data: { user: { bookings: [] } },
+      isLoading: false,
+      refetch: vi.fn(),
+    } as never);
+
+    render(
+      <MemoryRouter>
+        <EventBookingWizard
+          wizardOpen
+          section={{
+            id: "section-1",
+            name: "Events",
+            type: "EVENTS",
+            description: null,
+            purposeLinks: [
+              { purposes: ["ACCESS", "BOOKER"], userGroup: { id: "group-1", membershipStatuses: ["REGULAR"] } },
+            ],
+          } as never}
+          event={{
+            id: "event-1",
+            title: "Annual Dinner",
+            bookingStartDateTime: "2025-01-01T00:00:00Z",
+            bookingEndDateTime: "2030-01-01T00:00:00Z",
+            maxGuestsWithoutModeratorApproval: 1,
+            ticketTypes: [
+              {
+                id: "ticket-member",
+                title: "Member standard",
+                audience: TicketAudience.MEMBER,
+                price: 50,
+                userGroup: { id: "group-1", membershipStatuses: ["REGULAR"] },
+              },
+              {
+                id: "ticket-guest",
+                title: "Guest standard",
+                audience: TicketAudience.GUEST,
+                price: 25,
+                userGroup: { id: "group-1", membershipStatuses: ["REGULAR"] },
+              },
+            ],
+          } as never}
+        />
+      </MemoryRouter>
+    );
+
+    await user.click(screen.getByLabelText(/member standard/i));
+    await user.click(screen.getByRole("button", { name: "Next" }));
+
+    const guestCountInput = screen.getByLabelText(/how many guest tickets in total/i);
+    await user.clear(guestCountInput);
+    await user.type(guestCountInput, "4");
+    await user.click(screen.getByRole("button", { name: "Next" }));
+
+    const guestNames = screen.getAllByLabelText("Guest name");
+    await user.type(guestNames[0], "Guest One");
+    await user.type(guestNames[1], "Guest Two");
+    await user.type(guestNames[2], "Guest Three");
+    await user.type(guestNames[3], "Guest Four");
+    await user.click(screen.getByRole("button", { name: "Next" }));
+
+    expect(screen.getByText(/4 total \(3 via organiser review\)/i)).toBeInTheDocument();
+    expect(screen.getByRole("button", { name: "Confirm booking" })).toBeInTheDocument();
+    expect(screen.queryByRole("button", { name: "Next" })).not.toBeInTheDocument();
+  });
+
+  it("shows Next on review after returning from payment when booking is already submitted", async () => {
+    const user = userEvent.setup();
+    let bookings: Array<Record<string, unknown>> = [];
+    const submittedBooking = {
+      id: "booking-new",
+      status: BookingStatus.SUBMITTED,
+      revisionNumber: 1,
+      clientSubmissionKey: null,
+      bookerDietaryNote: "",
+      sitNextToUserIds: [],
+      accommodationRequested: false,
+      accommodationNote: null,
+      lines: [
+        {
+          id: "line-1",
+          sortOrder: 0,
+          guestDisplayName: null,
+          dietaryNote: null,
+          ticketType: {
+            id: "ticket-member",
+            title: "Member standard",
+            audience: TicketAudience.MEMBER,
+            price: 50,
+          },
+        },
+        {
+          id: "line-2",
+          sortOrder: 1,
+          guestDisplayName: "Guest One",
+          dietaryNote: null,
+          ticketType: {
+            id: "ticket-guest",
+            title: "Guest standard",
+            audience: TicketAudience.GUEST,
+            price: 25,
+          },
+        },
+      ],
+      guestTicketRequests: [
+        { id: "gtr-1", status: "PENDING", requestedGuestCount: 1 },
+        { id: "gtr-2", status: "PENDING", requestedGuestCount: 1 },
+        { id: "gtr-3", status: "PENDING", requestedGuestCount: 1 },
+      ],
+    };
+
+    vi.mocked(reactGenerated.useGetMyBookingsForEvent).mockImplementation(
+      () =>
+        ({
+          data: { user: { bookings } },
+          isLoading: false,
+          refetch: vi.fn().mockImplementation(async () => {
+            bookings = [submittedBooking];
+            return { data: { user: { bookings } } };
+          }),
+        }) as never
+    );
+
+    const wizardProps = {
+      wizardOpen: true,
+      section: {
+        id: "section-1",
+        name: "Events",
+        type: "EVENTS",
+        description: null,
+        purposeLinks: [
+          { purposes: ["ACCESS", "BOOKER"], userGroup: { id: "group-1", membershipStatuses: ["REGULAR"] } },
+        ],
+      },
+      event: {
+        id: "event-1",
+        title: "Annual Dinner",
+        bookingStartDateTime: "2025-01-01T00:00:00Z",
+        bookingEndDateTime: "2030-01-01T00:00:00Z",
+        maxGuestsWithoutModeratorApproval: 1,
+        ticketTypes: [
+          {
+            id: "ticket-member",
+            title: "Member standard",
+            audience: TicketAudience.MEMBER,
+            price: 50,
+            userGroup: { id: "group-1", membershipStatuses: ["REGULAR"] },
+          },
+          {
+            id: "ticket-guest",
+            title: "Guest standard",
+            audience: TicketAudience.GUEST,
+            price: 25,
+            userGroup: { id: "group-1", membershipStatuses: ["REGULAR"] },
+          },
+        ],
+      },
+    };
+
+    const { rerender } = render(
+      <MemoryRouter>
+        <EventBookingWizard {...(wizardProps as never)} />
+      </MemoryRouter>
+    );
+
+    await user.click(screen.getByLabelText(/member standard/i));
+    await user.click(screen.getByRole("button", { name: "Next" }));
+
+    const guestCountInput = screen.getByLabelText(/how many guest tickets in total/i);
+    await user.clear(guestCountInput);
+    await user.type(guestCountInput, "4");
+    await user.click(screen.getByRole("button", { name: "Next" }));
+
+    const guestNames = screen.getAllByLabelText("Guest name");
+    await user.type(guestNames[0], "Guest One");
+    await user.type(guestNames[1], "Guest Two");
+    await user.type(guestNames[2], "Guest Three");
+    await user.type(guestNames[3], "Guest Four");
+    await user.click(screen.getByRole("button", { name: "Next" }));
+    await user.click(screen.getByRole("button", { name: "Confirm booking" }));
+
+    await waitFor(() => {
+      expect(firebaseFunctions.submitEventBooking).toHaveBeenCalled();
+    });
+
+    rerender(
+      <MemoryRouter>
+        <EventBookingWizard {...(wizardProps as never)} />
+      </MemoryRouter>
+    );
+
+    await waitFor(() => {
+      expect(screen.getByText(/Pay for all tickets in one checkout/i)).toBeInTheDocument();
+    });
+
+    await user.click(screen.getByRole("button", { name: "Back" }));
+
+    expect(screen.getByText(/4 total \(3 via organiser review\)/i)).toBeInTheDocument();
+    expect(screen.getByRole("button", { name: "Next" })).toBeInTheDocument();
+    expect(screen.queryByRole("button", { name: "Confirm booking" })).not.toBeInTheDocument();
+
+    await user.click(screen.getByRole("button", { name: "Next" }));
+
+    expect(screen.getByText(/Pay for all tickets in one checkout/i)).toBeInTheDocument();
   });
 
   it("submits separate dietary notes for each guest", async () => {
