+++ src/features/sections/components/wizardSteps/PaymentStep.tsx
@@ -1,0 +1,96 @@
+import { Alert, Box, Button, Chip, Table, TableBody, TableCell, TableHead, TableRow, Typography } from "@mui/material";
+import { colors } from "../../../../config/colors";
+import { formatGbpMajorAmount } from "../../../../shared/utils/currencyDisplay";
+import { bookingTicketPaymentChipColor } from "../../utils/eventBookingStatusSummary";
+
+interface PaymentTicketRow {
+  id: string;
+  ticketTitle: string;
+  guestName?: string | null;
+  price: number;
+  paymentStatusLabel: string;
+  paymentStatus: string;
+  source: string;
+}
+
+interface PaymentStepProps {
+  paymentTicketRows: PaymentTicketRow[];
+  canProceedToConfirmation: boolean;
+  pendingGuestTicketsAwaitingApproval: boolean;
+  payingAllTickets: boolean;
+  onPayAllTickets: () => void;
+}
+
+export default function PaymentStep({
+  paymentTicketRows,
+  canProceedToConfirmation,
+  pendingGuestTicketsAwaitingApproval,
+  payingAllTickets,
+  onPayAllTickets,
+}: PaymentStepProps) {
+  return (
+    <Box>
+      <Typography variant="body2" color="text.secondary" sx={{ mb: 2 }}>
+        Pay for all tickets in one checkout to secure your place. You can also finish later and return
+        from your booking summary.
+      </Typography>
+
+      {paymentTicketRows.length > 0 ? (
+        <Table size="small" sx={{ mb: 2 }}>
+          <TableHead>
+            <TableRow>
+              <TableCell>Ticket</TableCell>
+              <TableCell>Guest</TableCell>
+              <TableCell>Price</TableCell>
+              <TableCell>Payment</TableCell>
+            </TableRow>
+          </TableHead>
+          <TableBody>
+            {paymentTicketRows.map((row) => (
+              <TableRow key={row.id}>
+                <TableCell>
+                  {row.ticketTitle}
+                  {row.source === "approved_guest_request" ? (
+                    <Typography component="span" variant="caption" color="text.secondary" sx={{ ml: 0.75 }}>
+                      (approved extra guest)
+                    </Typography>
+                  ) : null}
+                </TableCell>
+                <TableCell>{row.guestName ?? "—"}</TableCell>
+                <TableCell>{formatGbpMajorAmount(row.price)}</TableCell>
+                <TableCell>
+                  <Chip
+                    size="small"
+                    label={row.paymentStatusLabel}
+                    color={bookingTicketPaymentChipColor(row.paymentStatus)}
+                    variant={row.paymentStatus === "awaiting_approval" ? "outlined" : "filled"}
+                  />
+                </TableCell>
+              </TableRow>
+            ))}
+          </TableBody>
+        </Table>
+      ) : null}
+
+      {!canProceedToConfirmation ? (
+        <Button
+          variant="contained"
+          disabled={payingAllTickets}
+          onClick={onPayAllTickets}
+          sx={{ mt: 1, backgroundColor: colors.callToAction }}
+        >
+          {payingAllTickets ? "Starting checkout…" : "Pay for all tickets"}
+        </Button>
+      ) : pendingGuestTicketsAwaitingApproval ? (
+        <Alert severity="info" sx={{ mt: 1 }}>
+          All tickets due now are paid. When your additional guest request is approved, return here to
+          pay for their ticket.
+        </Alert>
+      ) : (
+        <Alert severity="success" sx={{ mt: 1 }}>
+          Payment complete. Continue to confirmation when you are ready.
+        </Alert>
+      )}
+    </Box>
+  );
+}
